cmake: do not exclusively set() in parent scope
when using lexy via FetchContent, cmake errors with the following:
> CMake Error (dev) at <...>/src/CMakeLists.txt:135 (target_sources):
>   uninitialized variable 'header_files'
> CMake Error (dev) at <...>/src/CMakeLists.txt:221 (target_sources):
>   uninitialized variable 'ext_headers_files'

This is because the set() calls at the top of the file previously passed
PARENT_SCOPE, resulting in these vars retaining an unset state.

Quoting from the cmake docs for command/set:
> If the PARENT_SCOPE option is given the variable will be set in the
> scope above the current scope. [...] The previous state of the
> variable's value stays the same in the current scope (e.g., if it was
> undefined before, it is still undefined and if it had a value, it is
> still that value)

Fix local uses of these vars by removing the PARENT_SCOPE flag when
calling set().

It's unclear to me what non-FetchContent build workflow (if any) would
break as a result of these variables being unset, so also backfill them
into the parent scope to retain the old behavior.

Signed-off-by: Nicholas Sielicki <opensource@nslick.com>